+++ NEWS.md
@@ -9,6 +9,8 @@ photobiologySensors 0.5.2
 
 * Data objects rebuilt with 'photobiology' 0.13.1.
 * Add data for the 14-channel AS7343 digital spectral sensor from ams-OSRAM.
+* Add data for the 3-channel AS7343 digital ultraviolet sensor from ams-OSRAM.
+* Add data for the TSL2591 digital ambient light sensor from ams-OSRAM.
 
 photobiologySensors 0.5.1
 =========================